more reorganization and desetoification
